% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\docType{package}
\name{aalmtools}
\alias{aalmtools-package}
\alias{aalmtools}
\title{aalmtools: Age-dependent Adult Lead Model Tools}
\description{
Implementation of the All-Ages Lead Model (AALM) in R with support for
parameter estimation, simulation, and visualization.
}
\section{Main Functions}{

\itemize{
  \item \code{\link{run_AALM}}: Run AALM simulation
  \item \code{\link{estimate_ingestion_rates}}: Bayesian estimation
  \item \code{\link{write_outputs_safe}}: Write model outputs
}
}

\author{
\strong{Maintainer}: Maria Cardelino \email{maria.cardelino@emory.edu}

}
