% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Actvty}
\alias{Actvty}
\title{Analytical solution for compartment activity
Direct port of Fortran Actvty function}
\usage{
Actvty(Y, P, X, D)
}
\arguments{
\item{Y}{Current amount}

\item{P}{Input rate}

\item{X}{Transfer coefficient}

\item{D}{Time step}
}
\description{
Analytical solution for compartment activity
Direct port of Fortran Actvty function
}
