% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_flows}
\alias{calculate_flows}
\title{Calculate flows between compartments}
\usage{
calculate_flows(state, sim, t)
}
\arguments{
\item{state}{Current model state}

\item{sim}{Simulation parameters}

\item{t}{Current timestep}
}
\value{
Updated flow rates
Calculate flows between compartments
}
\description{
Calculate flows between compartments
}
