% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_intake_uptake}
\alias{calculate_intake_uptake}
\title{Calculate intake and uptake values}
\usage{
calculate_intake_uptake(state, sim, t)
}
\arguments{
\item{state}{Current model state}

\item{sim}{Simulation parameters}

\item{t}{Current timestep}
}
\description{
Calculate intake and uptake values
}
