% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_mass_balance}
\alias{calculate_mass_balance}
\title{Calculate mass balance for all compartments}
\usage{
calculate_mass_balance(state, sim, t)
}
\arguments{
\item{state}{Current model state}

\item{sim}{Simulation parameters}

\item{t}{Current timestep}
}
\description{
Calculate mass balance for all compartments
}
