% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AALM.R
\name{initialize_arrays}
\alias{initialize_arrays}
\title{Initialize model arrays after parameters are read}
\usage{
initialize_arrays(state)
}
\arguments{
\item{state}{Model state}
}
\description{
Initialize model arrays after parameters are read
}
