% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{initialize_model}
\alias{initialize_model}
\title{Core simulation functions for AALM model
   State intitialization, mass balance calculations, 
   intake/update calculations (subroutine), results storage
Initialize model state}
\usage{
initialize_model(sim)
}
\arguments{
\item{sim}{Simulation parameters}
}
\value{
List containing model state vectors
}
\description{
Core simulation functions for AALM model
   State intitialization, mass balance calculations, 
   intake/update calculations (subroutine), results storage
Initialize model state
}
