% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_growth_params}
\alias{parse_growth_params}
\title{Parse growth parameters}
\usage{
parse_growth_params(line, sim)
}
\arguments{
\item{line}{Character string containing growth parameters}

\item{sim}{Simulation parameters list}
}
\description{
Parse growth parameters
}
