% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_lung_params}
\alias{parse_lung_params}
\title{Parse lung parameters}
\usage{
parse_lung_params(line, sim)
}
\arguments{
\item{line}{Character string containing lung parameters}

\item{sim}{Simulation parameters list}
}
\description{
Parse lung parameters
}
