% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_media_params}
\alias{parse_media_params}
\title{Parse media parameters}
\usage{
parse_media_params(line, sim)
}
\arguments{
\item{line}{Character string containing media parameters}

\item{sim}{Simulation parameters list}
}
\description{
Parse media parameters
}
