% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_phys_params}
\alias{parse_phys_params}
\title{Parse physiological parameters}
\usage{
parse_phys_params(line, PC)
}
\arguments{
\item{line}{Character string containing physiological parameters}

\item{PC}{Physiological constants vector}
}
\description{
Parse physiological parameters
}
