% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_run_name}
\alias{parse_run_name}
\title{Parse run name from input line}
\usage{
parse_run_name(line)
}
\arguments{
\item{line}{Character string containing the run name#'}
}
\value{
Parsed run name
}
\description{
Parse run name from input line
}
