% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_sim_params}
\alias{parse_sim_params}
\title{Parse simulation parameters from input file}
\usage{
parse_sim_params(line, sim)
}
\arguments{
\item{line}{Character string containing the parameter configuration}

\item{sim}{List object containing simulation settings to be updated}
}
\value{
Updated simulation settings list
}
\description{
Parses and validates simulation parameters from an input configuration line.
}
\examples{
\dontrun{
sim <- list()
line <- "OUTWRITE 1 0"
sim <- parse_sim_params(line, sim)
}
}
