% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_parameters}
\alias{read_parameters}
\title{Parameter file parsing functions for AALM model
   match read_sim_data(), read_growth_data(), etc in original Fortran model
Read and parse parameter file}
\usage{
read_parameters(filename, debug = FALSE)
}
\arguments{
\item{filename}{Path to parameter file}

\item{debug}{Print debug info}
}
\value{
List containing parsed parameters
}
\description{
Parameter file parsing functions for AALM model
   match read_sim_data(), read_growth_data(), etc in original Fortran model
Read and parse parameter file
}
