% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_AALM.R
\name{run_AALM}
\alias{run_AALM}
\title{Main entry point for AALM model
Direct port of Fortran PROGRAM Leg16}
\usage{
run_AALM(parFile = "LeggettInput.txt", outputDir = NULL, verbose = FALSE)
}
\arguments{
\item{parFile}{Path to parameter file}

\item{outputDir}{Output directory}

\item{verbose}{Print debug info

#' @return List containing model results:
\itemize{
  \item Y - Compartment amounts
  \item flows - Flow rates between compartments
  \item TS - Time series results
  \item MB - Mass balance information
}}
}
\description{
Main entry point for AALM model
Direct port of Fortran PROGRAM Leg16
}
\examples{
\dontrun{
results <- run_AALM("input/params.txt", "output", verbose=TRUE)
}

}
