% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_compartments}
\alias{update_compartments}
\title{Compartment update functions for AALM model with 
   Actvty() fuction
Update all compartments for current timestep}
\usage{
update_compartments(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point in simulation}
}
\value{
Updated state object
}
\description{
Compartment update functions for AALM model with 
   Actvty() fuction
Update all compartments for current timestep
}
