% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_cortical_diffuse}
\alias{update_cortical_diffuse}
\title{Update cortical bone diffuse}
\usage{
update_cortical_diffuse(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update cortical bone diffuse
}
