% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_cortical_volume}
\alias{update_cortical_volume}
\title{Update cortical volume compartment}
\usage{
update_cortical_volume(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update cortical volume compartment
}
