% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_feces}
\alias{update_feces}
\title{Update excretion compartment: feces}
\usage{
update_feces(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update excretion compartment: feces
}
