% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_hair}
\alias{update_hair}
\title{Update excretion compartment: hair}
\usage{
update_hair(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update excretion compartment: hair
}
