% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_kidney_2}
\alias{update_kidney_2}
\title{Update kidney compartment 2}
\usage{
update_kidney_2(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update kidney compartment 2
}
