% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_liver_1}
\alias{update_liver_1}
\title{Update liver compartment 1}
\usage{
update_liver_1(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update liver compartment 1
}
