% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_liver_2}
\alias{update_liver_2}
\title{Update liver 2 compartment}
\usage{
update_liver_2(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update liver 2 compartment
}
