% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_lower_large_intestine}
\alias{update_lower_large_intestine}
\title{Update lower large intestine compartment}
\usage{
update_lower_large_intestine(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update lower large intestine compartment
}
