% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_lung_alveolar}
\alias{update_lung_alveolar}
\title{Update lung alveolar region}
\usage{
update_lung_alveolar(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update lung alveolar region
}
