% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_lung_interstitial}
\alias{update_lung_interstitial}
\title{Update lung interstitial region}
\usage{
update_lung_interstitial(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update lung interstitial region
}
