% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_lung_TB}
\alias{update_lung_TB}
\title{Update lung compartment: TB}
\usage{
update_lung_TB(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update lung compartment: TB
}
