% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_plasma}
\alias{update_plasma}
\title{Update plasma compartment}
\usage{
update_plasma(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update plasma compartment
}
