% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_protein}
\alias{update_protein}
\title{Update protein compartment}
\usage{
update_protein(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update protein compartment
}
