% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_RBC}
\alias{update_RBC}
\title{Update RBC compartment}
\usage{
update_RBC(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update RBC compartment
}
