% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_small_intestine}
\alias{update_small_intestine}
\title{Update small intestine compartment}
\usage{
update_small_intestine(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update small intestine compartment
}
