% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_soft_tissue_2}
\alias{update_soft_tissue_2}
\title{Update soft tissue compartment 2}
\usage{
update_soft_tissue_2(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update soft tissue compartment 2
}
