% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_stomach}
\alias{update_stomach}
\title{Compartment updates ################
 1. Calculate outflows based on current compartment amount and transfer coefficients
 2. Update amount using Actvty()
 3. Store results in state vectors
Update stomach compartment}
\usage{
update_stomach(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Compartment updates ################
 1. Calculate outflows based on current compartment amount and transfer coefficients
 2. Update amount using Actvty()
 3. Store results in state vectors
Update stomach compartment
}
