% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_sweat}
\alias{update_sweat}
\title{Update excretion compartment: sweat}
\usage{
update_sweat(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update excretion compartment: sweat
}
