% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_trabecular_surface}
\alias{update_trabecular_surface}
\title{Update trabecular compartment: surface}
\usage{
update_trabecular_surface(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update trabecular compartment: surface
}
