% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_trabecular_volume}
\alias{update_trabecular_volume}
\title{Update trabecular compartment: volume}
\usage{
update_trabecular_volume(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update trabecular compartment: volume
}
