% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{update_urine}
\alias{update_urine}
\title{Update excretion compartment: urine}
\usage{
update_urine(state, t)
}
\arguments{
\item{state}{Current state of the AALM model}

\item{t}{Current time point}
}
\value{
Updated state
}
\description{
Update excretion compartment: urine
}
