% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate parsed parameters}
\usage{
validate_parameters(params)
}
\arguments{
\item{params}{List of parsed parameters}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Validate parsed parameters
}
