% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_concentrations_file}
\alias{write_concentrations_file}
\title{Write concentration outputs file}
\usage{
write_concentrations_file(state, sim, filename)
}
\arguments{
\item{state}{Model state}

\item{sim}{Simulation parameters}

\item{filename}{Output file path}
}
\description{
Write concentration outputs file
}
