% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_daily_file}
\alias{write_daily_file}
\title{Write daily inputs/outputs file}
\usage{
write_daily_file(state, sim, filename)
}
\arguments{
\item{state}{Model state}

\item{sim}{Simulation parameters}

\item{filename}{Output file path}
}
\description{
Write daily inputs/outputs file
}
