% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_log_file}
\alias{write_log_file}
\title{Write comprehensive log file}
\usage{
write_log_file(state, sim, filename)
}
\arguments{
\item{state}{Model state}

\item{sim}{Simulation parameters}

\item{filename}{Output file path}
}
\description{
Write comprehensive log file
}
