% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_outputs}
\alias{write_outputs}
\title{Output generation functions for AALM model
Direct port of Fortran output code}
\usage{
write_outputs(state, sim, runname, outputDir)
}
\arguments{
\item{state}{Final model state}

\item{sim}{Simulation parameters}

\item{runname}{Name of model run}

\item{outputDir}{Output directory}
}
\description{
Output generation functions for AALM model
Direct port of Fortran output code
}
