% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_source_file}
\alias{write_source_file}
\title{Write source tracking file}
\usage{
write_source_file(state, sim, filename)
}
\arguments{
\item{state}{Model state}

\item{sim}{Simulation parameters}

\item{filename}{Output file path}
}
\description{
Write source tracking file
}
